Rename: svg파일 assets/icons 폴더로 옮김
