--- src/components/common/Icons.tsx
@@ -1,107 +1,0 @@
-import React from 'react';
-
-const MoreVert = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-    <path
-      d="M14 6C14 7.10457 13.1046 8 12 8C10.8954 8 10 7.10457 10 6C10 4.89543 10.8954 4 12 4C13.1046 4 14 4.89543 14 6Z"
-      fill="#505050"
-    />
-    <path
-      d="M14 12C14 13.1046 13.1046 14 12 14C10.8954 14 10 13.1046 10 12C10 10.8954 10.8954 10 12 10C13.1046 10 14 10.8954 14 12Z"
-      fill="#505050"
-    />
-    <path
-      d="M12 20C13.1046 20 14 19.1046 14 18C14 16.8954 13.1046 16 12 16C10.8954 16 10 16.8954 10 18C10 19.1046 10.8954 20 12 20Z"
-      fill="#505050"
-    />
-  </svg>
-);
-
-const Profile = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="40" height="40" viewBox="0 0 40 40" fill="none">
-    <g clipPath="url(#clip0_18_21551)">
-      <circle cx="20" cy="20" r="19.375" fill="#F4F4F4" stroke="#CFCFCF" strokeWidth="1.25" />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M21.5093 21.0664C24.2317 20.3911 26.2495 17.9312 26.2495 15C26.2495 11.5482 23.4513 8.75 19.9995 8.75C16.5477 8.75 13.7495 11.5482 13.7495 15C13.7495 18.0042 15.8691 20.5133 18.6944 21.1135C11.061 21.9348 5.32842 27.4458 5.80503 33.9194C16.3983 43.5234 31.1457 37.2457 34.6561 33.2787C34.2127 27.2563 28.4862 21.6333 21.5093 21.0664Z"
-        fill="#CFCFCF"
-      />
-    </g>
-    <defs>
-      <clipPath id="clip0_18_21551">
-        <rect width="40" height="40" fill="white" />
-      </clipPath>
-    </defs>
-  </svg>
-);
-
-const Document = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-    <path
-      d="M8.5 8C8.22386 8 8 8.22386 8 8.5C8 8.77614 8.22386 9 8.5 9H11.5C11.7761 9 12 8.77614 12 8.5C12 8.22386 11.7761 8 11.5 8H8.5Z"
-      fill="#1D1D1D"
-    />
-    <path
-      d="M8 11.5C8 11.2239 8.22386 11 8.5 11H15.5C15.7761 11 16 11.2239 16 11.5C16 11.7761 15.7761 12 15.5 12H8.5C8.22386 12 8 11.7761 8 11.5Z"
-      fill="#1D1D1D"
-    />
-    <path
-      d="M8.5 14C8.22386 14 8 14.2239 8 14.5C8 14.7761 8.22386 15 8.5 15H15.5C15.7761 15 16 14.7761 16 14.5C16 14.2239 15.7761 14 15.5 14H8.5Z"
-      fill="#1D1D1D"
-    />
-    <path
-      d="M15.5858 4H6C5.44771 4 5 4.44772 5 5V19C5 19.5523 5.44772 20 6 20H18C18.5523 20 19 19.5523 19 19V7.41421C19 7.149 18.8946 6.89464 18.7071 6.70711L16.2929 4.29289C16.1054 4.10536 15.851 4 15.5858 4ZM6.5 5.5H14.25V8C14.25 8.41421 14.5858 8.75 15 8.75H17.5V18.5H6.5V5.5ZM15.75 5.87132L17.1287 7.25H15.75V5.87132Z"
-      fill="#1D1D1D"
-    />
-  </svg>
-);
-
-const Message = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-    <path
-      d="M10.5 8C10.2239 8 10 8.22386 10 8.5C10 8.77614 10.2239 9 10.5 9H13.5C13.7761 9 14 8.77614 14 8.5C14 8.22386 13.7761 8 13.5 8H10.5Z"
-      fill="black"
-    />
-    <path
-      d="M10.5 13C10.2239 13 10 13.2239 10 13.5C10 13.7761 10.2239 14 10.5 14H13.5C13.7761 14 14 13.7761 14 13.5C14 13.2239 13.7761 13 13.5 13H10.5Z"
-      fill="black"
-    />
-    <path
-      d="M9 10.5C8.72386 10.5 8.5 10.7239 8.5 11C8.5 11.2761 8.72386 11.5 9 11.5H15C15.2761 11.5 15.5 11.2761 15.5 11C15.5 10.7239 15.2761 10.5 15 10.5H9Z"
-      fill="black"
-    />
-    <path
-      d="M10.0911 17.8181L5.97353 19.5828C5.58518 19.7492 5.17827 19.3888 5.29657 18.9832L6.19741 15.8946C6.20807 15.858 6.19695 15.8186 6.16908 15.7926C4.824 14.5402 4 12.8543 4 11C4 7.13401 7.58172 4 12 4C16.4183 4 20 7.13401 20 11C20 14.866 16.4183 18 12 18C11.3636 18 10.7446 17.935 10.1511 17.8121C10.1309 17.8079 10.11 17.81 10.0911 17.8181ZM12 16.5C15.7843 16.5 18.5 13.8557 18.5 11C18.5 8.14431 15.7843 5.5 12 5.5C8.21574 5.5 5.5 8.14431 5.5 11C5.5 12.3906 6.11447 13.6922 7.19126 14.6948C7.62816 15.1017 7.80904 15.7261 7.63741 16.3146L7.3297 17.3696L9.50019 16.4394C9.80357 16.3094 10.1371 16.2774 10.4552 16.3433C10.9487 16.4454 11.4659 16.5 12 16.5Z"
-      fill="black"
-    />
-  </svg>
-);
-
-const Edit = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-    <path
-      d="M17.8083 3.21373C18.0933 2.92876 18.5553 2.92876 18.8403 3.21373L20.7863 5.15968C21.0712 5.44465 21.0712 5.90669 20.7863 6.19167L12.0295 14.9484C11.9741 15.0039 11.91 15.05 11.8399 15.0851L7.94798 17.0311C7.66704 17.1715 7.32774 17.1165 7.10564 16.8944C6.88354 16.6723 6.82848 16.333 6.96895 16.052L8.91489 12.1601C8.94997 12.09 8.99612 12.0259 9.05158 11.9705L17.8083 3.21373Z"
-      fill="#1D1D1D"
-    />
-    <path
-      d="M3.72973 3.97297C3.32671 3.97297 3 4.29968 3 4.7027V20.2703C3 20.6733 3.32671 21 3.72973 21H19.2973C19.7003 21 20.027 20.6733 20.027 20.2703V13.4594C20.027 13.0564 19.7003 12.7297 19.2973 12.7297C18.8943 12.7297 18.5675 13.0564 18.5675 13.4594V19.5405H4.45946V5.43243H10.5405C10.9435 5.43243 11.2703 5.10572 11.2703 4.7027C11.2703 4.29968 10.9435 3.97297 10.5405 3.97297H3.72973Z"
-      fill="#1D1D1D"
-    />
-  </svg>
-);
-
-const Filter = () => {
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-    <path
-      d="M8 11C9.30622 11 10.4175 10.1652 10.8293 9H20C20.5523 9 21 8.55228 21 8C21 7.44772 20.5523 7 20 7H10.8293C10.4175 5.83481 9.30622 5 8 5C6.69378 5 5.58254 5.83481 5.17071 7H4C3.44772 7 3 7.44772 3 8C3 8.55228 3.44772 9 4 9H5.17071C5.58254 10.1652 6.69378 11 8 11ZM8 9C7.44772 9 7 8.55228 7 8C7 7.44772 7.44772 7 8 7C8.55228 7 9 7.44772 9 8C9 8.55228 8.55228 9 8 9Z"
-      fill="black"
-    />
-    <path
-      d="M16 13C14.6938 13 13.5825 13.8348 13.1707 15L4 15C3.44772 15 3 15.4477 3 16C3 16.5523 3.44772 17 4 17L13.1707 17C13.5825 18.1652 14.6938 19 16 19C17.3062 19 18.4175 18.1652 18.8293 17H20C20.5523 17 21 16.5523 21 16C21 15.4477 20.5523 15 20 15H18.8293C18.4175 13.8348 17.3062 13 16 13ZM16 15C16.5523 15 17 15.4477 17 16C17 16.5523 16.5523 17 16 17C15.4477 17 15 16.5523 15 16C15 15.4477 15.4477 15 16 15Z"
-      fill="black"
-    />
-  </svg>;
-};
-
-export { MoreVert, Profile, Document, Message, Edit, Filter };
